+++ CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.5.0] - 2018-05-29
+### Changed
+- Updated `VerifyClient` to use the JSON endpoints instead of XML.
+- Updated endpoints which are used by `VerifyClient` from public to package scope in order to encourage usage through `VerifyClient`.
+- Deprecated XML version of the following endpoints:
+    - `VerifyEndpoint`
+    - `CheckEndpoint`
+    - `SearchEndpoint`
+   
+- Deprecated the following XML results:
+    - `VerifyResult` should use `VerifyResponse`
+    - `CheckResult` should use `CheckResponse`
+    - `SearchResult` should use `SearchVerifyResponse`
+    
+- Deprecated the following XML methods:
+    - `VerifyCheckMethod`
+
+### Added
+- Added `VerifyStatus` enumeration to use for statuses coming back from the verify endpoint.
+- Added `VerifyResponse`, `CheckResponse`, and `SearchVerifyResponse` for JSON responses to match other JSON using endpoints.
+- Added `VerifyMethod`, `CheckMethod`, and `SearchMethod` for better segregation between endpoint and method classes.
+- Add `split` attribute to the `RecordNcco` object.
+
+### Fixed
+- Updated `ConversationNcco`'s `musicOnHoldUrl` to serialize into an array for use in the Voice API.
+
 ## [3.4.1] - 2018-05-11
 
 ### Changed